Port lua to use ui.xml now X4 7.50 can load lua properly
